services/svc-hrms-core/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.1.2
python-dotenv
azure-storage-blob